infra/requirements.txt
aws-cdk-lib>=2.100.0,<3.0
constructs>=10.0.0,<11.0
